Made a RoomAssigner which assigns rooms to a square floor plan.
Tested with up 6x6 floor plan.  Rooms are 10x10 with like 1 unit walls, so realistical walking space is like 8x8.  Easy to scale.  I could change the code to get the prefab length and width instead of hardcoding the 10 in, at a later date.  The rooms are correctly being randomized also.